# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse:
    properties:
      error:
        example: Mensaje de error legible
        type: string
    type: object
...
  github_com_johansgiraldo_noteops_backend_internal_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_johansgiraldo_noteops_backend_internal_models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.User'
    type: object
...
  github_com_johansgiraldo_noteops_backend_internal_models.RecordGradeRequest:
    properties:
      activity_id:
        type: string
      comment:
        type: string
      enrollment_id:
        type: string
      value:
        maximum: 5
        minimum: 0
        type: number
    required:
    - activity_id
    - enrollment_id
    - value
    type: object
...
  github_com_johansgiraldo_noteops_backend_internal_models.UpdateCommentRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
...
host: localhost:8080
info:
  contact:
    name: Johan Sebastian Giraldo Hurtado
  description: API REST de gestión de notas académicas con cálculo automático de nota
    definitiva, sesiones de clase en tiempo real y reserva de turnos.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: NoteOPs API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Valida credenciales y devuelve un JWT válido por 24 horas.
      parameters:
      - description: Correo y contraseña
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      summary: Autenticar usuario
      tags:
      - auth
  /grades:
    post:
      consumes:
      - application/json
      description: Crea o actualiza (upsert) la nota de una actividad para una matrícula.
      parameters:
      - description: Datos de la nota (valor entre 0 y 5)
        in: body
        name: grade
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.RecordGradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Grade'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar o actualizar una nota
      tags:
      - grades
  /grades/{id}/comment:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID de la nota (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Nuevo comentario
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar el comentario de una nota
      tags:
      - grades
...
  /sessions:
    post:
      consumes:
      - application/json
      description: Crea una sesión y genera automáticamente los espacios (slots) según
        la duración.
      parameters:
      - description: Datos de la sesión
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.CreateSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear sesión de clase con sus espacios
      tags:
      - sessions
  /sessions/{id}/activate:
    post:
      parameters:
      - description: ID de la sesión (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Iniciar una sesión de clase
      tags:
      - sessions
  /sessions/{id}/deactivate:
    post:
      parameters:
      - description: ID de la sesión (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finalizar una sesión de clase
      tags:
      - sessions
  /sessions/{id}/slots:
    get:
      parameters:
      - description: ID de la sesión (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Slot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      summary: Listar los espacios (slots) de una sesión
      tags:
      - slots
  /sessions/{id}/slots/{slotID}/reserve:
    post:
      consumes:
      - application/json
      description: Reserva un turno para un estudiante. Devuelve 409 si el espacio
        ya está tomado.
      parameters:
      - description: ID de la sesión (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: ID del espacio (UUID)
        in: path
        name: slotID
        required: true
        type: string
      - description: ID del estudiante
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ReserveSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Slot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      summary: Reservar un espacio de una sesión
      tags:
      - slots
  /sessions/active:
    get:
      parameters:
      - description: ID de la materia (UUID)
        in: query
        name: subject_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      summary: Obtener la sesión activa de una materia
      tags:
      - sessions
  /students:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del estudiante
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.RegisterStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear estudiante
      tags:
      - students
...
  /subjects/{id}/students:
    get:
      parameters:
      - description: ID de la materia (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.Student'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_johansgiraldo_noteops_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar estudiantes de una materia
      tags:
      - subjects
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Escribe **Bearer &lt;token&gt;** usando el JWT obtenido en /api/auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"